Refractor

+++ WareHouseManagement/Model/Receipt/CustomerBuyReceiptDetail.cs
@@ -7,8 +7,8 @@ namespace WareHouseManagement.Model.Receipt
         public string ProductId { get; set; }
         public string ReceiptId { get; set; }
         public int Quantity { get; set; }
-        public int PriceOfOne {  get; set; }
-        public int TotalPrice { get; set; }
+        public float PriceOfOne {  get; set; }
+        public float TotalPrice { get; set; }
         public virtual Product ProductNav { get; set; }
         public virtual CustomerBuyReceipt ReceiptNav { get; set; }
